(PC-29799)[PRO] feat: Deleted Offerer page.
