+++ pro/src/pages/Home/Offerers/OffererDetails.tsx
@@ -4,11 +4,10 @@ import { useDispatch, useSelector } from 'react-redux'
 import { useNavigate, useSearchParams } from 'react-router-dom'
 
 import { useAnalytics } from 'app/App/analytics/firebase'
-import { Events, OffererLinkEvents } from 'core/FirebaseEvents/constants'
+import { OffererLinkEvents } from 'core/FirebaseEvents/constants'
 import { SAVED_OFFERER_ID_KEY } from 'core/shared/constants'
 import { SelectOption } from 'custom_types/form'
 import fullAddUserIcon from 'icons/full-add-user.svg'
-import fullEditIcon from 'icons/full-edit.svg'
 import { updateSelectedOffererId } from 'store/user/reducer'
 import { selectCurrentOffererId } from 'store/user/selectors'
 import { ButtonLink } from 'ui-kit/Button/ButtonLink'
@@ -76,7 +75,7 @@ export const OffererDetails = ({
           <ButtonLink
             variant={ButtonVariant.TERNARY}
             link={{
-              to: `/structures/${selectedOffererId}`,
+              to: `/collaborateurs`,
               isExternal: false,
             }}
             icon={fullAddUserIcon}
@@ -88,25 +87,6 @@ export const OffererDetails = ({
             }}
           >
             Inviter
-          </ButtonLink>
-
-          <div className={cn(styles['separator'])} />
-
-          <ButtonLink
-            variant={ButtonVariant.TERNARY}
-            link={{
-              to: `/structures/${String(selectedOffererId)}`,
-              isExternal: false,
-            }}
-            icon={fullEditIcon}
-            isDisabled={!isUserOffererValidated}
-            onClick={() =>
-              logEvent(Events.CLICKED_MODIFY_OFFERER, {
-                offerer_id: selectedOffererId ?? undefined,
-              })
-            }
-          >
-            Modifier
           </ButtonLink>
         </div>
       </div>
