--- pro/src/pages/Offerers/Offerer/OffererDetails/ApiKey/ApiKey.module.scss
@@ -1,79 +1,0 @@
-@use "styles/mixins/_fonts.scss" as fonts;
-@use "styles/mixins/_rem.scss" as rem;
-
-.api-key {
-  .title {
-    @include fonts.title4;
-
-    margin-top: rem.torem(24px);
-
-    .counter-error {
-      color: var(--color-error-light);
-    }
-  }
-
-  .info {
-    margin-top: rem.torem(8px);
-  }
-
-  .list {
-    display: flex;
-    flex-direction: column;
-    margin-top: rem.torem(24px);
-
-    .item {
-      align-items: center;
-      background-color: var(--color-grey-light);
-      border-radius: rem.torem(4px);
-      display: flex;
-      justify-content: space-between;
-      padding: rem.torem(12px) rem.torem(16px);
-      width: 100%;
-
-      &:not(:first-child) {
-        margin-top: rem.torem(8px);
-      }
-
-      .text {
-        color: var(--color-grey-dark);
-        word-break: break-all;
-
-        &--new-key {
-          color: var(--color-primary);
-
-          @include fonts.button;
-        }
-      }
-
-      .action {
-        margin-left: rem.torem(16px);
-      }
-    }
-  }
-
-  .generate {
-    margin-top: rem.torem(24px);
-    min-width: rem.torem(180px);
-  }
-}
-
-.api-key-dialog {
-  align-items: center;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  max-width: rem.torem(532px);
-  text-align: center;
-}
-
-.section {
-  margin-bottom: rem.torem(32px);
-
-  .main-list-title {
-    @include fonts.title3;
-
-    border-bottom: rem.torem(2px) solid var(--color-black);
-    padding-bottom: rem.torem(8px);
-    margin-bottom: rem.torem(24px);
-  }
-}
