--- pro/src/pages/Offerers/Offerer/OffererDetails/OffererDetails.module.scss
@@ -1,71 +1,0 @@
-@use "styles/mixins/_fonts.scss" as fonts;
-@use "styles/mixins/_rem.scss" as rem;
-
-.offerer-form-heading {
-  margin-bottom: rem.torem(32px);
-}
-
-.title-page {
-  display: flex;
-  justify-content: space-between;
-}
-
-.offerer-name {
-  margin-bottom: rem.torem(8px);
-}
-
-.description {
-  @include fonts.body;
-
-  margin-top: rem.torem(20px);
-}
-
-.offerer-page-container {
-  max-width: rem.torem(486px);
-}
-
-.offerer-page {
-  position: relative;
-
-  .has-text-grey {
-    color: var(--color-grey-dark);
-  }
-
-  .op-detail {
-    margin: rem.torem(8px) 0;
-
-    > :first-child {
-      color: var(--color-grey-dark);
-      margin-right: rem.torem(5px);
-    }
-  }
-
-  .op-subtitle {
-    @include fonts.body;
-  }
-
-  .offerer-page-go-back-link {
-    margin-bottom: rem.torem(30px);
-    color: var(--color-primary);
-  }
-}
-
-.section {
-  margin-bottom: rem.torem(32px);
-  padding-bottom: 0;
-  padding-top: 0;
-
-  .main-list-title {
-    @include fonts.title3;
-
-    border-bottom: rem.torem(2px) solid var(--color-black);
-    margin-bottom: rem.torem(24px);
-    padding-bottom: rem.torem(8px);
-  }
-
-  .title-actions-container {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-  }
-}
