--- pro/src/pages/Offerers/Offerer/OffererDetails/Venues/VenueItem/VenueItem.module.scss
@@ -1,31 +1,0 @@
-@use "styles/mixins/_rem.scss" as rem;
-@use "styles/mixins/_fonts.scss" as fonts;
-
-.venue-item {
-  display: flex;
-  width: 100%;
-
-  &:not(:last-child) {
-		margin-bottom: rem.torem(24px);
-	}
-}
-
-.name {
-  @include fonts.title3;
-  
-  margin-bottom: rem.torem(8px);
-}
-
-.address {
-  text-transform: uppercase;
-}
-
-.picto {
-  height: rem.torem(20px);
-  width: rem.torem(20px);
-  margin-right: rem.torem(16px);
-}
-
-.create-offer-button {
-  margin-top: rem.torem(8px);
-}
