--- pro/src/pages/Offerers/Offerer/OffererDetails/Venues/Venues.module.scss
@@ -1,26 +1,0 @@
-@use "styles/mixins/_fonts.scss" as fonts;
-@use "styles/mixins/_rem.scss" as rem;
-
-.section {
-  margin-bottom: rem.torem(8px);
-}
-
-.main-list-title {
-  border-bottom: rem.torem(2px) solid var(--color-black);
-  margin-bottom: rem.torem(24px);
-  padding-bottom: rem.torem(8px);
-  display: inline-flex;
-  width: 100%;
-
-  & > h2 {
-    @include fonts.title3;
-  }
-
-  & > a {
-    margin-left: auto;
-  }
-}
-
-.venue-list {
-  padding-bottom: rem.torem(24px);
-}
