--- pro/src/pages/VenueCreation/__specs__/venueSubmitRedirectUrl.spec.tsx
@@ -1,24 +1,0 @@
-import { SharedCurrentUserResponseModel } from 'apiClient/v1'
-import { defaultGetOffererResponseModel } from 'utils/individualApiFactories'
-
-import { venueSubmitRedirectUrl } from '../venueSubmitRedirectUrl'
-
-const offerer = { ...defaultGetOffererResponseModel }
-
-describe('redirect url after submit', () => {
-  it('should redirect admin user to the right url', () => {
-    const url = venueSubmitRedirectUrl(offerer.id, {
-      isAdmin: true,
-    } as SharedCurrentUserResponseModel)
-
-    expect(url).toEqual('/structures/1')
-  })
-
-  it('redirect non admin user to the right url', () => {
-    const url = venueSubmitRedirectUrl(offerer.id, {
-      isAdmin: false,
-    } as SharedCurrentUserResponseModel)
-
-    expect(url).toEqual('/accueil?success')
-  })
-})
